Split report creation (+tests) into multiple files.
