+++ lib/pana.dart
@@ -2,13 +2,13 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-export 'src/create_report.dart' show renderSimpleSectionSummary;
 export 'src/dartdoc_analyzer.dart';
 export 'src/download_utils.dart';
 export 'src/maintenance.dart';
 export 'src/model.dart';
 export 'src/package_analyzer.dart';
 export 'src/pubspec.dart';
+export 'src/report/create_report.dart' show renderSimpleSectionSummary;
 export 'src/sdk_env.dart';
 export 'src/utils.dart' show runProc, yamlToJson;
 export 'src/version.dart';
